[animals] added simple APIs for cats, dogs and catfacts
