+++ api/__init__.py
@@ -9,4 +9,4 @@ from .translate import translate
 from .tts import tts
 from .ud import ud
 from .goodreads import goodreads
-from .shiba import shiba
+from .animals import shiba, fox, cat, catfact
